scale image module done
